+++ app/Filament/Resources/Posts/Pages/ViewPost.php
@@ -19,7 +19,7 @@ class ViewPost extends ViewRecord
     {
         parent::mount($record);
 
-        $this->record->increment('views');
+        $this->record::withoutTimestamps(fn () => $this->record->increment('views'));
     }
 
     protected function getHeaderActions(): array
@@ -38,7 +38,7 @@ class ViewPost extends ViewRecord
                     $pin->exists() ? $pin->delete() : Pin::create([
                         'user_id' => Auth::id(),
                         'pinnable_type' => $this->record->getMorphClass(),
-                        'pinnable_id' => $this->record->id
+                        'pinnable_id' => $this->record->id,
                     ]);
                 })
                 ->visible(fn (): bool => Auth::check()),