+++ app/Filament/Widgets/MostViewed.php
@@ -8,8 +8,6 @@ use Filament\Widgets\Widget;
 
 class MostViewed extends Widget
 {
-    protected static bool $isLazy = false;
-
     protected int | string | array $columnSpan = 'full';
 
     protected string $view = 'filament.widgets.most-viewed';